+++ src/whitelist_filter.cpp
@@ -31,11 +31,11 @@
 
 #include "whitelist_filter.h"
 
-void swd::whitelist_filter::set_id(const unsigned int& id) {
+void swd::whitelist_filter::set_id(const unsigned long long& id) {
     id_ = id;
 }
 
-unsigned int swd::whitelist_filter::get_id() const {
+unsigned long long swd::whitelist_filter::get_id() const {
     return id_;
 }
